Adicionando validações de token em mutations

+++ Graphql/src/graphql/user/resolver.js
@@ -1,13 +1,14 @@
 import { AuthenticationError } from "apollo-server";
+import { checkUserIsLogged } from "../post/resolver";
 
 export const user = async (_, { id }, { dataSources }) =>
-  dataSources.userApi.getUser(id);
+  await dataSources.userApi.getUser(id);
 
 export const users = async (_, { input }, { dataSources }) =>
-  dataSources.userApi.getUsers(input);
+  await dataSources.userApi.getUsers(input);
 
 export const createUser = async (_, { data }, { dataSources }) =>
-  dataSources.userApi.createUser(data);
+  await dataSources.userApi.createUser(data);
 
 export const updateUser = async (
   _,
@@ -22,5 +23,14 @@ export const updateUser = async (
   return dataSources.userApi.updateUser(id, data);
 };
 
-export const deleteUser = async (_, { id }, { dataSources }) =>
-  dataSources.userApi.deleteUser(id);
+export const deleteUser = async (_, { id }, { dataSources, loggedUserId }) => {
+  checkUserIsLogged(id);
+  checkOwner(id, loggedUserId);
+
+  return dataSources.userApi.deleteUser(id);
+};
+
+function checkOwner(userId, loggedUserId) {
+  if (loggedUserId !== userId)
+    throw new AuthenticationError("You cannot update other user");
+}
